fixed "New prompt"
